fix(security): implement operator security review remediation (O1-O5) + audit policy
Implement all findings from the operator security review:

- O1: Fix EVM source-chain verification routing — verify_deposit_on_source
  now routes by src_chain_id via source_chain_endpoints map; unknown
  source chains fail closed (refuse to approve)
- O2: Add bearer token auth — /status and /pending gated by optional
  OPERATOR_API_TOKEN; /health and /metrics remain public
- O3: Harden HTTP parsing — structured request-line extraction replaces
  substring matching; connection concurrency bounded by semaphore (256);
  5s read timeout; both api.rs and metrics.rs hardened
- O4: Upgrade vulnerable deps — sqlx 0.7→0.8.6, bytes 1.11.0→1.11.1,
  protobuf removed (prometheus default-features=false for operator and
  canceler); 3 unfixable transitive advisories documented in audit.toml
- O5: Replace unsafe static mut — OnceLock<Instant> replaces
  static mut START_TIME; all unsafe blocks removed from api.rs

Additional:
- Add .cargo/audit.toml for operator, canceler, multichain-rs with
  documented ignore list for unfixable transitive advisories
- Add .githooks/pre-commit hook running cargo audit on staged packages
- Fix 18 pre-existing clippy warnings in multichain-rs (derivable_impls,
  type_complexity, too_many_arguments, unnecessary casts)
- Format canceler and multichain-rs (cargo fmt)

+++ packages/multichain-rs/src/evm/watcher.rs
@@ -562,7 +562,7 @@ pub fn parse_deposit_log(log: &alloy::rpc::types::Log) -> Option<DepositEvent> {
 
     let block_number = log.block_number?;
     let tx_hash = log.transaction_hash?;
-    let log_index = log.log_index? as u64;
+    let log_index = log.log_index?;
 
     // topic[1] = destChain (bytes4, left-padded to bytes32)
     let dest_chain_bytes = topics[1];
@@ -637,7 +637,7 @@ pub fn parse_withdraw_submit_log(log: &alloy::rpc::types::Log) -> Option<Withdra
 
     let block_number = log.block_number?;
     let tx_hash = log.transaction_hash?;
-    let log_index = log.log_index? as u64;
+    let log_index = log.log_index?;
 
     // topic[1] = withdrawHash (bytes32)
     let withdraw_hash = topics[1];
@@ -693,7 +693,7 @@ pub fn parse_withdraw_approve_log(log: &alloy::rpc::types::Log) -> Option<Withdr
 
     let block_number = log.block_number?;
     let tx_hash = log.transaction_hash?;
-    let log_index = log.log_index? as u64;
+    let log_index = log.log_index?;
 
     let withdraw_hash = topics[1];
 
@@ -716,7 +716,7 @@ pub fn parse_withdraw_cancel_log(log: &alloy::rpc::types::Log) -> Option<Withdra
 
     let block_number = log.block_number?;
     let tx_hash = log.transaction_hash?;
-    let log_index = log.log_index? as u64;
+    let log_index = log.log_index?;
 
     let withdraw_hash = topics[1];
 
@@ -748,7 +748,7 @@ pub fn parse_withdraw_execute_log(log: &alloy::rpc::types::Log) -> Option<Withdr
 
     let block_number = log.block_number?;
     let tx_hash = log.transaction_hash?;
-    let log_index = log.log_index? as u64;
+    let log_index = log.log_index?;
 
     let withdraw_hash = topics[1];
